disable virtualenvwrapper
virtualenvwrapper+oh-my-zsh/plugins/virtualenvwrapper+nix{,OS} do not
mix well ATM.

virtualenwrapper is a shell file that is expected to be sourced. NixPkgs
wrappes this script which, when sourced, result in clobbering the
environment.

The wrapper does something like this:

export PATH=...
export VIRTUALENVWRAPPER_PYTHONPATH=...
source $wrapped

This results in a login shell containing /nix/store/...python/bin and
the like in the PATH.

I noticed this when I was trying to "python -c 'import sqlite3'" which
would through errors, even though python27Full was installed. It turned
out that there were multiple python binaries in the environment, with
the /nix/store/... ones -- which are the non-Full version -- having
higher precedence.